Addding AbstractCloudBigtableTableDoFn for complex user CBT DoFns. Multiple users contacted us about DoFns that require some additional Gets, including issue #607. This PR should help.
